+++ react/components/Button/index.js
@@ -23,6 +23,7 @@ class Button extends Component {
     const {
       size,
       block,
+      fullHeight,
       variation,
       icon,
       children,
@@ -123,7 +124,11 @@ class Button extends Component {
     }
 
     if (block) {
-      classes += 'w-100 h-100 '
+      classes += 'w-100 '
+    }
+
+    if (fullHeight) {
+      classes += 'h-100 '
     }
 
     return (
@@ -166,6 +171,7 @@ class Button extends Component {
 Button.defaultProps = {
   size: 'regular',
   block: false,
+  fullHeight: false,
   variation: 'primary',
   disabled: false,
   autoFocus: false,
@@ -187,6 +193,8 @@ Button.propTypes = {
   ]),
   /** Block style */
   block: PropTypes.bool,
+  /** Height 100% */
+  fullHeight: PropTypes.bool,
   /** Loading state */
   isLoading: PropTypes.bool,
   /** [DEPRECATED] If you are using just an Icon component inside, use this as true */